// src/lib/event/types.ts
export enum Key {
  ENTER = 'Enter',
  BACKSPACE = 'Backspace',

  Q = 'q',
  W = 'w',
  E = 'e',
  R = 'r',
  T = 't',
  Y = 'y',
  U = 'u',
  I = 'i',
  O = 'o',
  P = 'p',

  A = 'a',
  S = 's',
  D = 'd',
  F = 'f',
  G = 'g',
  H = 'h',
  J = 'j',
  K = 'k',
  L = 'l',

  Z = 'z',
  X = 'x',
  C = 'c',
  V = 'v',
  B = 'b',
  N = 'n',
  M = 'm',
}